src/main/java/org/sagebionetworks/bridge/hibernate/TemplatePersistenceExceptionConverter.java
package org.sagebionetworks.bridge.hibernate;

import javax.persistence.OptimisticLockException;
import javax.persistence.PersistenceException;

import org.springframework.stereotype.Component;

import org.sagebionetworks.bridge.exceptions.ConcurrentModificationException;

@Component
public class TemplatePersistenceExceptionConverter implements PersistenceExceptionConverter {

    @Override
    public RuntimeException convert(PersistenceException exception, Object entity) {
        if (exception instanceof OptimisticLockException) {
            return new ConcurrentModificationException((HibernateTemplate)entity);
        }
        // there are no constraints that would prevent deletion at this point.
        return exception;        
    }

}
